refactor!: Make ThreadLocalMap internal & move mtc package into transactions
- ThreadLocalMap is only used as a private property of an internal class, which
is only used in transaction methods.
- There is no use case for this class to be part of the public api, or in its own package.
