+++ exposed-r2dbc/api/exposed-r2dbc.api
@@ -608,11 +608,6 @@ public final class org/jetbrains/exposed/v1/r2dbc/mappers/ValueTypeMapper : org/
 	public fun setValue (Lio/r2dbc/spi/Statement;Lorg/jetbrains/exposed/v1/core/vendors/DatabaseDialect;Lorg/jetbrains/exposed/v1/r2dbc/mappers/R2dbcTypeMapping;Lorg/jetbrains/exposed/v1/core/IColumnType;Ljava/lang/Object;I)Z
 }
 
-public final class org/jetbrains/exposed/v1/r2dbc/mtc/ThreadLocalMap : java/lang/InheritableThreadLocal {
-	public fun <init> ()V
-	public synthetic fun childValue (Ljava/lang/Object;)Ljava/lang/Object;
-}
-
 public class org/jetbrains/exposed/v1/r2dbc/statements/BatchInsertSuspendExecutable : org/jetbrains/exposed/v1/r2dbc/statements/InsertSuspendExecutable {
 	public fun <init> (Lorg/jetbrains/exposed/v1/core/statements/BatchInsertStatement;)V
 	public fun getStatement ()Lorg/jetbrains/exposed/v1/core/statements/BatchInsertStatement;
