+++ exposed-r2dbc/src/main/kotlin/org/jetbrains/exposed/v1/r2dbc/transactions/mtc/MappedTransactionContext.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.exposed.v1.r2dbc.mtc
+package org.jetbrains.exposed.v1.r2dbc.transactions.mtc
 
 import org.jetbrains.exposed.v1.r2dbc.R2dbcTransaction
